Fix parsing of bounds in summary stats
Previously, bounds were assumed to be in the default unit system
(MeV, MeV/c, m), whereas the rest of the input file was assumed
to be in SI. This inconsistency has been fixed, which means that
energy and momentum bounds in the stats block need, e.g.,
factors of MeV, GeV, MeV/c etc, for earlier results to be
reproduced. A warning will be issued for such bounds by
at least the next release of Ptarmigan.